pyproject.toml
[project]
name = "Palindrome-Data-AI-Solutions-Engineer"
version = "0.1.0"
description = "Risk assessment system for HIV and mental health using LLM and hybrid approaches"
readme = "README.md"
requires-python = ">=3.11"


dependencies = [
    "anthropic>=0.74.1",
    "chromadb>=1.3.5",
    "faiss-cpu>=1.13.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "pypdf>=6.4.0",
    "pypdf2>=3.0.1",
    "python-dateutil>=2.9.0.post0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]


[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
